backend/requirements.txt
apscheduler==3.11.0
atproto==0.0.61
django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
instagrapi==2.1.3
minio==7.2.15
psycopg2-binary==2.9.10
tweepy==4.15.0
pillow